Change input type from float to int for summation

+++ sum_numbers.py
@@ -1,9 +1,10 @@
 # Program to sum two numbers
 
 # Method 1: Using input from user
-num1 = float(input("Enter first number: "))
-num2 = float(input("Enter second number: "))
+num1 = int(input("Enter first number: "))
+num2 = int(input("Enter second number: "))
 
 sum = num1 + num2
 
 print(f"The sum of {num1} and {num2} is {sum}")
+
